app/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv==1.0.1
gunicorn
rich
openai
numpy
scikit-learn
soundfile
websocket-client
pypdf
PyPDF2
python-docx
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]
passlib[bcrypt]
aiohttp
aiohttp-socks
pytest-asyncio